+++ fidimag/atomistic/llg.py
@@ -31,7 +31,7 @@ class LLG(AtomisticDriver):
                  name,
                  data_saver,
                  use_jac,
-                 integrator='sundials'
+                 integrator='cvode_bdf'
                  ):
 
         # Inherit from the driver class