+++ fidimag/atomistic/llg_stt.py
@@ -47,7 +47,7 @@ class LLG_STT(AtomisticDriver):
                  name,
                  data_saver,
                  use_jac,
-                 integrator='sundials'
+                 integrator='cvode_bdf'
                  ):
 
         # Inherit from the driver class