+++ fidimag/micro/llg.py
@@ -30,7 +30,7 @@ class LLG(MicroDriver):
                  interactions,
                  name,
                  data_saver,
-                 integrator='sundials',
+                 integrator='cvode_bdf',
                  use_jac=False
                  ):