+++ fidimag/micro/sim.py
@@ -62,7 +62,7 @@ class Sim(SimBase):
 
     def __init__(self, mesh, name='unnamed', driver='llg',
                  # Integrator arguments:
-                 integrator='sundials', use_jac=False
+                 integrator='cvode_bdf', use_jac=False
                  ):
 
         super().__init__(mesh, name)